+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Specifying the spacy version to 2.2.4 in `setup.py`
 
 ## [v0.0.7](https://github.com/danieldeutsch/qaeval/releases/tag/0.0.7) - 2021-05-07
 ### Added