+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added trying to fix the predicted character offsets using an alignment algorithm
+- Added an option to return the QA result as a dict
 
 ### Fixed
 - Specifying the spacy version to 2.2.4 in `setup.py`